chore: Refactor GIF creation logic in video.js
This commit refactors the GIF creation logic in the video.js file. It introduces a new function, createGIFObject, to create the GIF object with the specified settings. The addFramesToGIF function is updated to handle adding frames to the GIF object based on the number of images provided. The setupGIFEvents function is added to handle progress and finished events of the GIF rendering. Additionally, a timeout is set to reject the GIF creation promise if it takes longer than 20 minutes.